feat: respect CSS page size
Browsers let us use CSS rules to specify what page size should be used
by default when printing (or when printing to a PDF as we do). Chrome
needs a flag to be set to respect these CSS rules.

Implementing it like this lets us move all the styling concerns to the
CSS, which we can later bundle alongside the HTML into an easier to use
template, instead of needing to set the flags in the API call (you can
manually set the page size and height in inches).

This could also let us support multiple page sizes down the line if we
ever need that.